working on getting code completion working
